@@ -100,33 +76,45 @@ ignore_missing_imports = true
 
 
 [tool.ruff]
-target-version = "py311"
+target-version = "py313"
 line-length = 88
+extend-exclude = ["tests/initial_test_state"]
 
 [tool.ruff.lint]
 select = [
-    "E",  # pycodestyle errors
-    "W",  # pycodestyle warnings
-    "F",  # pyflakes
-    "I",  # isort
-    "B",  # flake8-bugbear
-    "C4", # flake8-comprehensions
-    "UP", # pyupgrade
-    "RUF100", # unused noqa
+    "E",     # pycodestyle errors
+    "W",     # pycodestyle warnings
+    "F",     # pyflakes
+    "I",     # isort
+    "B",     # flake8-bugbear
+    "C4",    # flake8-comprehensions
+    "UP",    # pyupgrade
+    "RUF",   # ruff-specific rules
+    "PIE",   # misc lints
+    "PERF",  # performance anti-patterns
+    "ASYNC", # async best practices
+    "A",     # builtin shadowing
+    "LOG",   # logging best practices
+    "SIM",   # simplify
 ]
 ignore = [
-    "E501",  # line too long, handled by black
-    "B008",  # do not perform function calls in argument defaults
-    "C901",  # too complex
-    "B904",  # raise exceptions with from None/err - non-critical
-    "F841",  # local variable assigned but never used
-    "B007",  # loop control variable not used within loop body
-    "W293",  # blank line contains whitespace
-    "W291",  # trailing whitespace
-    "F821",  # undefined name (often false positives in complex imports)
-    "E722",  # bare except (temporary for cleanup)
-    "UP038", # isinstance union syntax (non-critical)
-    "I001",  # import sorting handled by isort
+    "E501",    # line too long — formatter handles this
+    "B008",    # function calls in defaults — FastMCP pattern
+    "C901",    # complexity — too many to fix now
+    "SIM102",  # collapsible-if — sometimes less readable
+    "SIM108",  # ternary — sometimes less readable
+    "SIM105",  # contextlib.suppress — style preference
+    "ASYNC109",# timeout params are standard in HA API patterns
+    "RUF001",  # ambiguous unicode — HA entity names use degree symbols etc
+    "RUF003",  # ambiguous unicode in comments
+    "RUF010",  # explicit f-string type conversion — str(x) is clearer than !s
+    "F841",    # unused variable — too many to fix now
+    "RUF059",  # unused unpacked variable
+    "SIM118",  # in-dict-keys — style preference
+    "PIE810",  # multiple-starts-ends-with — style preference
+    "RUF005",  # collection-literal-concatenation
+    "RUF022",  # unsorted-dunder-all
+    "RUF013",  # implicit-optional
 ]
 
 [tool.ruff.lint.per-file-ignores]
