@@ -4,6 +4,7 @@ Smart search tools for Home Assistant MCP server.
 
 import asyncio
 import logging
+import time
 from typing import Any
 
 from ..client.rest_client import HomeAssistantClient
@@ -13,7 +14,16 @@ from ..utils.fuzzy_search import calculate_partial_ratio, create_fuzzy_searcher
 logger = logging.getLogger(__name__)
 
 # Default concurrency limit for parallel operations
-DEFAULT_CONCURRENCY_LIMIT = 5
+DEFAULT_CONCURRENCY_LIMIT = 20
+
+# Bulk fetch timeouts (in seconds)
+BULK_REST_TIMEOUT = 5.0  # Timeout for bulk REST endpoint calls
+BULK_WEBSOCKET_TIMEOUT = 3.0  # Timeout for bulk WebSocket calls
+INDIVIDUAL_CONFIG_TIMEOUT = 5.0  # Timeout for individual config fetches
+
+# Time budgets for fallback individual fetching (in seconds)
+AUTOMATION_CONFIG_TIME_BUDGET = 15.0  # Max time for fetching automation configs individually
+SCRIPT_CONFIG_TIME_BUDGET = 10.0  # Max time for fetching script configs individually
 
 
 class SmartSearchTools:
@@ -36,7 +46,12 @@ class SmartSearchTools:
         self.fuzzy_searcher = create_fuzzy_searcher(threshold=fuzzy_threshold)
 
     async def smart_entity_search(
-        self, query: str, limit: int = 10, include_attributes: bool = False, domain_filter: str | None = None
+        self,
+        query: str,
+        limit: int = 10,
+        offset: int = 0,
+        include_attributes: bool = False,
+        domain_filter: str | None = None,
     ) -> dict[str, Any]:
         """
         Advanced entity search with fuzzy matching and typo tolerance.
@@ -44,6 +59,7 @@ class SmartSearchTools:
         Args:
             query: Search query (can be partial, with typos)
             limit: Maximum number of results
+            offset: Number of results to skip for pagination
             include_attributes: Whether to include full entity attributes
             domain_filter: Optional domain to filter entities before search (e.g., "light", "sensor")
 
@@ -58,12 +74,15 @@ class SmartSearchTools:
             # This ensures fuzzy search only looks at entities in the target domain
             if domain_filter:
                 entities = [
-                    e for e in entities
+                    e
+                    for e in entities
                     if e.get("entity_id", "").startswith(f"{domain_filter}.")
                 ]
 
-            # Perform fuzzy search - returns (limited_results, total_count)
-            matches, total_matches = self.fuzzy_searcher.search_entities(entities, query, limit)
+            # Perform fuzzy search - returns (paginated_results, total_count)
+            matches, total_matches = self.fuzzy_searcher.search_entities(
+                entities, query, limit, offset
+            )
 
             # Format results
             results = []
@@ -95,33 +114,22 @@ class SmartSearchTools:
 
                 results.append(result)
 
-            # Get suggestions if no good matches
-            suggestions = []
-            if not matches or (matches and matches[0]["score"] < 80):
-                suggestions = self.fuzzy_searcher.get_smart_suggestions(entities, query)
+            has_more = (offset + len(results)) < total_matches
 
-            response = {
+            response: dict[str, Any] = {
                 "success": True,
                 "query": query,
                 "total_matches": total_matches,
-                "matches": results,  # Changed from 'results' to 'matches' for consistency
-                "is_truncated": total_matches > len(results),
-                "search_metadata": {
-                    "fuzzy_threshold": self.settings.fuzzy_threshold,
-                    "best_match_score": matches[0]["score"] if matches else 0,
-                    "search_suggestions": suggestions,
-                },
-                "usage_tips": [
-                    "Try partial names: 'living' finds 'Living Room Light'",
-                    "Domain search: 'light' finds all light entities",
-                    "French/English: 'salon' or 'living' both work",
-                    "Typo tolerant: 'lihgt' finds 'light' entities",
-                ],
+                "offset": offset,
+                "limit": limit,
+                "count": len(results),
+                "has_more": has_more,
+                "next_offset": offset + limit if has_more else None,
+                "matches": results,
             }
 
-            # Add hint if results are truncated
-            if response["is_truncated"]:
-                response["usage_tips"].insert(0, f"Showing {len(results)} of {total_matches} matches. Increase 'limit' parameter to see more.")
+            if not matches or (matches and matches[0]["score"] < 80):
+                response["suggestions"] = self.fuzzy_searcher.get_smart_suggestions(entities, query)
 
             return response
 
@@ -213,11 +221,16 @@ class SmartSearchTools:
             for area_id, area_info in area_registry.items():
                 area_name = area_info.get("name", "")
                 # Exact match on area_id or name (case-insensitive)
-                if area_query_lower == area_id.lower() or area_query_lower == area_name.lower():
+                if (
+                    area_query_lower == area_id.lower()
+                    or area_query_lower == area_name.lower()
+                ):
                     matched_area_ids.add(area_id)
                     continue
                 # Fuzzy match on area name
-                name_score = calculate_partial_ratio(area_query_lower, area_name.lower())
+                name_score = calculate_partial_ratio(
+                    area_query_lower, area_name.lower()
+                )
                 id_score = calculate_partial_ratio(area_query_lower, area_id.lower())
                 best_score = max(name_score, id_score)
                 if best_score >= 80:
@@ -229,18 +242,9 @@ class SmartSearchTools:
                     "total_areas_found": 0,
                     "total_entities": 0,
                     "areas": {},
-                    "search_metadata": {
-                        "grouped_by_domain": group_by_domain,
-                        "area_inference_method": "registry_lookup",
-                        "available_areas": [
-                            {"area_id": aid, "name": ainfo.get("name", aid)}
-                            for aid, ainfo in area_registry.items()
-                        ],
-                    },
-                    "usage_tips": [
-                        "No areas matched your query. See available_areas for valid area names.",
-                        "Try exact area names from Home Assistant",
-                        "Use ha_list_areas() to see all areas",
+                    "available_areas": [
+                        {"area_id": aid, "name": ainfo.get("name", aid)}
+                        for aid, ainfo in area_registry.items()
                     ],
                 }
 
@@ -249,8 +253,9 @@ class SmartSearchTools:
             entity_area_resolved: dict[str, str] = {}
             for entity_id, reg_info in entity_reg_map.items():
                 area_id = reg_info.get("area_id")
-                if not area_id and reg_info.get("device_id"):
-                    area_id = device_area_map.get(reg_info["device_id"])
+                device_id = reg_info.get("device_id")
+                if not area_id and device_id:
+                    area_id = device_area_map.get(device_id)
                 if area_id:
                     entity_area_resolved[entity_id] = area_id
 
@@ -304,7 +309,9 @@ class SmartSearchTools:
                     area_data["entities"] = [
                         {
                             "entity_id": entity_id,
-                            "friendly_name": (state_info := state_map.get(entity_id, {}))
+                            "friendly_name": (
+                                state_info := state_map.get(entity_id, {})
+                            )
                             .get("attributes", {})
                             .get("friendly_name", entity_id),
                             "domain": entity_id.split(".")[0],
@@ -321,15 +328,6 @@ class SmartSearchTools:
                 "total_areas_found": len(formatted_areas),
                 "total_entities": total_entities,
                 "areas": formatted_areas,
-                "search_metadata": {
-                    "grouped_by_domain": group_by_domain,
-                    "area_inference_method": "registry_lookup",
-                },
-                "usage_tips": [
-                    "Try room names: 'salon', 'chambre', 'cuisine'",
-                    "English names: 'living', 'bedroom', 'kitchen'",
-                    "Use ha_list_areas() to see all available areas",
-                ],
             }
 
         except Exception as e:
@@ -500,12 +498,12 @@ class SmartSearchTools:
                 "most_common_domains": [domain for domain, _ in sorted_domains[:5]],
                 "controllable_devices": [
                     domain
-                    for domain in domain_stats.keys()
+                    for domain in domain_stats
                     if domain in ["light", "switch", "climate", "media_player", "cover"]
                 ],
                 "monitoring_sensors": [
                     domain
-                    for domain in domain_stats.keys()
+                    for domain in domain_stats
                     if domain in ["sensor", "binary_sensor", "camera"]
                 ],
                 "automation_ready": "automation" in domain_stats
@@ -520,10 +518,15 @@ class SmartSearchTools:
                 all_entities = stats["all_entities"]
 
                 # Apply max_entities_per_domain limit
-                if max_entities_per_domain and len(all_entities) > max_entities_per_domain:
+                if (
+                    max_entities_per_domain
+                    and len(all_entities) > max_entities_per_domain
+                ):
                     # Random selection for minimal
                     if detail_level == "minimal":
-                        selected_entities = random.sample(all_entities, max_entities_per_domain)
+                        selected_entities = random.sample(
+                            all_entities, max_entities_per_domain
+                        )
                     else:
                         # Take first N for other levels
                         selected_entities = all_entities[:max_entities_per_domain]
@@ -580,7 +583,9 @@ class SmartSearchTools:
         self,
         query: str,
         search_types: list[str] | None = None,
-        limit: int = 20,
+        limit: int = 5,
+        offset: int = 0,
+        include_config: bool = False,
         concurrency_limit: int = DEFAULT_CONCURRENCY_LIMIT,
     ) -> dict[str, Any]:
         """
@@ -592,8 +597,10 @@ class SmartSearchTools:
         Args:
             query: Search query (can be partial, with typos)
             search_types: Types to search (default: ["automation", "script", "helper"])
-            limit: Maximum total results to return
-            concurrency_limit: Max concurrent API calls for config fetching (default: 5)
+            limit: Maximum total results to return (default: 5)
+            offset: Number of results to skip for pagination (default: 0)
+            include_config: Include full config in results (default: False)
+            concurrency_limit: Max concurrent API calls for config fetching
 
         Returns:
             Dictionary with search results grouped by type
@@ -613,130 +620,268 @@ class SmartSearchTools:
             # Fetch all entities once at the beginning to avoid repeated calls
             all_entities = await self.client.get_states()
 
+            # Pre-resolve unique_ids from cached entity states to avoid redundant API calls
+            automation_unique_id_map = {}
+            for e in all_entities:
+                eid = e.get("entity_id", "")
+                if eid.startswith("automation."):
+                    uid = e.get("attributes", {}).get("id")
+                    if uid:
+                        automation_unique_id_map[eid] = uid
+
             # Create semaphore for limiting concurrent API calls
             semaphore = asyncio.Semaphore(concurrency_limit)
 
-            # Search automations with parallel config fetching
+            # ================================================================
+            # AUTOMATION SEARCH
+            # Uses a 3-tier strategy to fetch configs within the MCP timeout:
+            #   A) Try REST bulk endpoint (single call for all configs)
+            #   B) Try WebSocket bulk endpoints
+            #   C) Fall back to individual REST calls with a time budget,
+            #      prioritizing automations that best match the query by name
+            # ================================================================
             if "automation" in search_types:
                 automation_entities = [
-                    e for e in all_entities if e.get("entity_id", "").startswith("automation.")
+                    e
+                    for e in all_entities
+                    if e.get("entity_id", "").startswith("automation.")
                 ]
 
-                async def fetch_automation_config(entity: dict[str, Any]) -> dict[str, Any] | None:
-                    """Fetch automation config with semaphore-controlled concurrency."""
+                # Phase 1: Score ALL automations by name (instant, no API calls)
+                name_scored: list[tuple[str, str, int, str | None]] = []
+                for entity in automation_entities:
                     entity_id = entity.get("entity_id", "")
-                    friendly_name = entity.get("attributes", {}).get("friendly_name", entity_id)
-
-                    # Check if query matches in name first
-                    name_match_score = self.fuzzy_searcher._calculate_entity_score(
+                    friendly_name = entity.get("attributes", {}).get(
+                        "friendly_name", entity_id
+                    )
+                    name_score = self.fuzzy_searcher._calculate_entity_score(
                         entity_id, friendly_name, "automation", query_lower
                     )
-
-                    # Get automation config with concurrency control
-                    async with semaphore:
-                        try:
-                            config_response = await self.client.get_automation_config(entity_id)
-                            config_match_score = self._search_in_dict(config_response, query_lower)
-
-                            total_score = max(name_match_score, config_match_score)
-
-                            if total_score >= self.settings.fuzzy_threshold:
-                                return {
-                                    "entity_id": entity_id,
-                                    "friendly_name": friendly_name,
-                                    "score": total_score,
-                                    "match_in_name": name_match_score >= self.settings.fuzzy_threshold,
-                                    "match_in_config": config_match_score >= self.settings.fuzzy_threshold,
-                                    "config": config_response,
-                                }
-                        except Exception as e:
-                            logger.debug(f"Could not get config for {entity_id}: {e}")
-                            # Still include if name matches
-                            if name_match_score >= self.settings.fuzzy_threshold:
-                                return {
-                                    "entity_id": entity_id,
-                                    "friendly_name": friendly_name,
-                                    "score": name_match_score,
-                                    "match_in_name": True,
-                                    "match_in_config": False,
-                                }
-                    return None
-
-                # Fetch all automation configs in parallel
-                automation_results = await asyncio.gather(
-                    *[fetch_automation_config(e) for e in automation_entities],
-                    return_exceptions=True,
-                )
-
-                # Filter out None results and exceptions
-                for result in automation_results:
-                    if isinstance(result, dict):
-                        results["automations"].append(result)
-                    elif isinstance(result, Exception):
-                        logger.debug(f"Automation config fetch failed: {result}")
-
-            # Search scripts with parallel config fetching
-            if "script" in search_types:
-                script_entities = [
-                    e for e in all_entities if e.get("entity_id", "").startswith("script.")
-                ]
-
-                async def fetch_script_config(entity: dict[str, Any]) -> dict[str, Any] | None:
-                    """Fetch script config with semaphore-controlled concurrency."""
-                    entity_id = entity.get("entity_id", "")
-                    friendly_name = entity.get("attributes", {}).get("friendly_name", entity_id)
-                    script_id = entity_id.replace("script.", "")
-
-                    # Check if query matches in name
-                    name_match_score = self.fuzzy_searcher._calculate_entity_score(
-                        entity_id, friendly_name, "script", query_lower
+                    unique_id = automation_unique_id_map.get(entity_id)
+                    name_scored.append(
+                        (entity_id, friendly_name, name_score, unique_id)
                     )
 
-                    # Get script config with concurrency control
-                    async with semaphore:
+                # Phase 2: Try to bulk-fetch ALL automation configs with a single API call
+                all_automation_configs: dict[str, dict[str, Any]] = {}
+                bulk_fetched = False
+
+                # Attempt A: REST bulk endpoint /config/automation/config (no ID)
+                try:
+                    resp = await asyncio.wait_for(
+                        self.client._request("GET", "/config/automation/config"),
+                        timeout=BULK_REST_TIMEOUT,
+                    )
+                    if isinstance(resp, list):
+                        for item in resp:
+                            uid = item.get("id")
+                            if uid:
+                                all_automation_configs[uid] = item
+                        bulk_fetched = True
+                except Exception as e:
+                    logger.debug(f"Automation REST bulk fetch failed: {e}")
+
+                # Attempt B: WebSocket bulk endpoints
+                if not bulk_fetched:
+                    for ws_type in [
+                        "config/automation/config/list",
+                        "automation/config/list",
+                    ]:
+                        if bulk_fetched:
+                            break
                         try:
-                            config_response = await self.client.get_script_config(script_id)
-                            script_config = config_response.get("config", {})
-                            config_match_score = self._search_in_dict(script_config, query_lower)
-
-                            total_score = max(name_match_score, config_match_score)
-
-                            if total_score >= self.settings.fuzzy_threshold:
-                                return {
-                                    "entity_id": entity_id,
-                                    "script_id": script_id,
-                                    "friendly_name": friendly_name,
-                                    "score": total_score,
-                                    "match_in_name": name_match_score >= self.settings.fuzzy_threshold,
-                                    "match_in_config": config_match_score >= self.settings.fuzzy_threshold,
-                                    "config": script_config,
-                                }
+                            ws_resp = await asyncio.wait_for(
+                                self.client.send_websocket_message({"type": ws_type}),
+                                timeout=BULK_WEBSOCKET_TIMEOUT,
+                            )
+                            if isinstance(ws_resp, dict) and ws_resp.get("success"):
+                                for item in ws_resp.get("result", []):
+                                    uid = item.get("id")
+                                    if uid:
+                                        all_automation_configs[uid] = item
+                                bulk_fetched = True
                         except Exception as e:
-                            logger.debug(f"Could not get config for {script_id}: {e}")
-                            # Still include if name matches
-                            if name_match_score >= self.settings.fuzzy_threshold:
-                                return {
-                                    "entity_id": entity_id,
-                                    "script_id": script_id,
-                                    "friendly_name": friendly_name,
-                                    "score": name_match_score,
-                                    "match_in_name": True,
-                                    "match_in_config": False,
-                                }
-                    return None
+                            logger.debug(f"Automation WebSocket bulk fetch ({ws_type}) failed: {e}")
 
-                # Fetch all script configs in parallel
-                script_results = await asyncio.gather(
-                    *[fetch_script_config(e) for e in script_entities],
-                    return_exceptions=True,
-                )
+                # Attempt C: Individual REST calls with time budget (LAST RESORT)
+                # Prioritize name-matched automations so we at least get their configs
+                if not bulk_fetched:
+                    budget_start = time.perf_counter()
+                    sorted_by_score = sorted(
+                        name_scored, key=lambda x: x[2], reverse=True
+                    )
 
-                # Filter out None results and exceptions
-                for result in script_results:
-                    if isinstance(result, dict):
-                        results["scripts"].append(result)
-                    elif isinstance(result, Exception):
-                        logger.debug(f"Script config fetch failed: {result}")
+                    for (
+                        _entity_id,
+                        _friendly_name,
+                        _name_score,
+                        unique_id,
+                    ) in sorted_by_score:
+                        if time.perf_counter() - budget_start > AUTOMATION_CONFIG_TIME_BUDGET:
+                            break
+                        if not unique_id or unique_id in all_automation_configs:
+                            continue
+                        try:
+                            config = await asyncio.wait_for(
+                                self.client._request(
+                                    "GET", f"/config/automation/config/{unique_id}"
+                                ),
+                                timeout=INDIVIDUAL_CONFIG_TIMEOUT,
+                            )
+                            all_automation_configs[unique_id] = config
+                        except Exception as e:
+                            logger.debug(f"Automation individual config fetch ({unique_id}) failed: {e}")
+
+                # Phase 3: Score with whatever configs we have
+                for entity_id, friendly_name, name_score, unique_id in name_scored:
+                    config = (
+                        all_automation_configs.get(unique_id, {}) if unique_id else {}
+                    )
+                    config_match_score = (
+                        self._search_in_dict(config, query_lower) if config else 0
+                    )
+                    total_score = max(name_score, config_match_score)
+
+                    if total_score >= self.settings.fuzzy_threshold:
+                        results["automations"].append(
+                            {
+                                "entity_id": entity_id,
+                                "friendly_name": friendly_name,
+                                "score": total_score,
+                                "match_in_name": name_score
+                                >= self.settings.fuzzy_threshold,
+                                "match_in_config": config_match_score
+                                >= self.settings.fuzzy_threshold,
+                                "config": config if config else None,
+                            }
+                        )
+
+            # ================================================================
+            # SCRIPT SEARCH (same 3-tier strategy: REST bulk -> WS bulk -> individual)
+            # ================================================================
+            if "script" in search_types:
+                script_entities = [
+                    e
+                    for e in all_entities
+                    if e.get("entity_id", "").startswith("script.")
+                ]
+
+                # Phase 1: Score all scripts by name (instant)
+                script_name_scored: list[tuple[str, str, str, int]] = []
+                for entity in script_entities:
+                    entity_id = entity.get("entity_id", "")
+                    friendly_name = entity.get("attributes", {}).get(
+                        "friendly_name", entity_id
+                    )
+                    script_id = entity_id.replace("script.", "")
+                    name_score = self.fuzzy_searcher._calculate_entity_score(
+                        entity_id, friendly_name, "script", query_lower
+                    )
+                    script_name_scored.append(
+                        (entity_id, friendly_name, script_id, name_score)
+                    )
+
+                # Phase 2: Try bulk fetch for scripts
+                all_script_configs: dict[str, dict[str, Any]] = {}
+                script_bulk_fetched = False
+
+                # Attempt A: REST bulk endpoint
+                try:
+                    resp = await asyncio.wait_for(
+                        self.client._request("GET", "/config/script/config"),
+                        timeout=INDIVIDUAL_CONFIG_TIMEOUT,
+                    )
+                    if isinstance(resp, list):
+                        for item in resp:
+                            sid = item.get("id") or item.get(
+                                "alias", ""
+                            ).lower().replace(" ", "_")
+                            if sid:
+                                all_script_configs[sid] = item
+                        script_bulk_fetched = True
+                except Exception as e:
+                    logger.debug(f"Script REST bulk fetch failed: {e}")
+
+                # Attempt B: WebSocket bulk endpoints
+                if not script_bulk_fetched:
+                    for ws_type in [
+                        "config/script/config/list",
+                        "script/config/list",
+                    ]:
+                        if script_bulk_fetched:
+                            break
+                        try:
+                            ws_resp = await asyncio.wait_for(
+                                self.client.send_websocket_message({"type": ws_type}),
+                                timeout=BULK_WEBSOCKET_TIMEOUT,
+                            )
+                            if isinstance(ws_resp, dict) and ws_resp.get("success"):
+                                for item in ws_resp.get("result", []):
+                                    sid = item.get("id") or item.get(
+                                        "alias", ""
+                                    ).lower().replace(" ", "_")
+                                    if sid:
+                                        all_script_configs[sid] = item
+                                script_bulk_fetched = True
+                        except Exception as e:
+                            logger.debug(f"Script WebSocket bulk fetch ({ws_type}) failed: {e}")
+
+                # Attempt C: Individual fetch with budget
+                if not script_bulk_fetched:
+                    budget_start = time.perf_counter()
+                    sorted_scripts = sorted(
+                        script_name_scored, key=lambda x: x[3], reverse=True
+                    )
+                    for (
+                        _entity_id,
+                        _friendly_name,
+                        script_id,
+                        _name_score,
+                    ) in sorted_scripts:
+                        if time.perf_counter() - budget_start > SCRIPT_CONFIG_TIME_BUDGET:
+                            break
+                        if script_id in all_script_configs:
+                            continue
+                        try:
+                            config_resp = await asyncio.wait_for(
+                                self.client.get_script_config(script_id),
+                                timeout=INDIVIDUAL_CONFIG_TIMEOUT,
+                            )
+                            all_script_configs[script_id] = config_resp.get(
+                                "config", {}
+                            )
+                        except Exception as e:
+                            logger.debug(f"Script individual config fetch ({script_id}) failed: {e}")
+
+                # Phase 3: Score scripts
+                for (
+                    entity_id,
+                    friendly_name,
+                    script_id,
+                    name_score,
+                ) in script_name_scored:
+                    script_config = all_script_configs.get(script_id, {})
+                    config_match_score = (
+                        self._search_in_dict(script_config, query_lower)
+                        if script_config
+                        else 0
+                    )
+                    total_score = max(name_score, config_match_score)
+
+                    if total_score >= self.settings.fuzzy_threshold:
+                        results["scripts"].append(
+                            {
+                                "entity_id": entity_id,
+                                "script_id": script_id,
+                                "friendly_name": friendly_name,
+                                "score": total_score,
+                                "match_in_name": name_score
+                                >= self.settings.fuzzy_threshold,
+                                "match_in_config": config_match_score
+                                >= self.settings.fuzzy_threshold,
+                                "config": script_config if script_config else None,
+                            }
+                        )
 
             # Search helpers with parallel WebSocket calls
             if "helper" in search_types:
@@ -754,7 +899,9 @@ class SmartSearchTools:
                     async with semaphore:
                         try:
                             message = {"type": f"{helper_type}/list"}
-                            helper_list_response = await self.client.send_websocket_message(message)
+                            helper_list_response = (
+                                await self.client.send_websocket_message(message)
+                            )
 
                             if not helper_list_response.get("success"):
                                 return []
@@ -768,23 +915,31 @@ class SmartSearchTools:
                                 name = helper.get("name", helper_id)
 
                                 # Check if query matches in name or config
-                                name_match_score = self.fuzzy_searcher._calculate_entity_score(
-                                    entity_id, name, helper_type, query_lower
+                                name_match_score = (
+                                    self.fuzzy_searcher._calculate_entity_score(
+                                        entity_id, name, helper_type, query_lower
+                                    )
                                 )
-                                config_match_score = self._search_in_dict(helper, query_lower)
+                                config_match_score = self._search_in_dict(
+                                    helper, query_lower
+                                )
 
                                 total_score = max(name_match_score, config_match_score)
 
                                 if total_score >= self.settings.fuzzy_threshold:
-                                    helper_results.append({
-                                        "entity_id": entity_id,
-                                        "helper_type": helper_type,
-                                        "name": name,
-                                        "score": total_score,
-                                        "match_in_name": name_match_score >= self.settings.fuzzy_threshold,
-                                        "match_in_config": config_match_score >= self.settings.fuzzy_threshold,
-                                        "config": helper,
-                                    })
+                                    helper_results.append(
+                                        {
+                                            "entity_id": entity_id,
+                                            "helper_type": helper_type,
+                                            "name": name,
+                                            "score": total_score,
+                                            "match_in_name": name_match_score
+                                            >= self.settings.fuzzy_threshold,
+                                            "match_in_config": config_match_score
+                                            >= self.settings.fuzzy_threshold,
+                                            "config": helper,
+                                        }
+                                    )
 
                             return helper_results
                         except Exception as e:
@@ -804,48 +959,42 @@ class SmartSearchTools:
                     elif isinstance(result, Exception):
                         logger.debug(f"Helper list fetch failed: {result}")
 
-            # Sort all results by score and apply limit
-            all_results = []
-            for result_type, items in results.items():
-                for item in items:
-                    item["result_type"] = result_type.rstrip("s")  # singular form
-                    all_results.append(item)
+            # Merge all results with their category, sort by score, and paginate
+            tagged_results: list[tuple[str, dict[str, Any]]] = []
+            for category, items in results.items():
+                tagged_results.extend((category, item) for item in items)
 
-            all_results.sort(key=lambda x: x["score"], reverse=True)
-            limited_results = all_results[:limit]
+            tagged_results.sort(key=lambda x: x[1]["score"], reverse=True)
 
-            # Re-group by type
+            total_before_pagination = len(tagged_results)
+            paginated = tagged_results[offset:offset + limit]
+
+            # Re-group paginated results by category
             final_results: dict[str, list[dict[str, Any]]] = {
                 "automations": [],
                 "scripts": [],
                 "helpers": [],
             }
-            for item in limited_results:
-                result_type = item.pop("result_type")
-                final_results[f"{result_type}s"].append(item)
+            for category, item in paginated:
+                if not include_config:
+                    item.pop("config", None)
+                final_results[category].append(item)
 
-            total_matches = len(limited_results)
+            has_more = (offset + len(paginated)) < total_before_pagination
 
-            # Return automations/scripts/helpers at top level for easy access
             return {
                 "success": True,
                 "query": query,
-                "total_matches": total_matches,
+                "total_matches": total_before_pagination,
+                "offset": offset,
+                "limit": limit,
+                "count": len(paginated),
+                "has_more": has_more,
+                "next_offset": offset + limit if has_more else None,
                 "automations": final_results["automations"],
                 "scripts": final_results["scripts"],
                 "helpers": final_results["helpers"],
                 "search_types": search_types,
-                "search_metadata": {
-                    "fuzzy_threshold": self.settings.fuzzy_threshold,
-                    "best_match_score": limited_results[0]["score"] if limited_results else 0,
-                    "truncated": len(all_results) > limit,
-                },
-                "usage_tips": [
-                    "Deep search finds matches in automation triggers, actions, and conditions",
-                    "Script sequences and service calls are also searched",
-                    "Helper configurations including options and constraints are included",
-                    "Use match_in_name and match_in_config to understand where the match occurred",
-                ],
             }
 
         except Exception as e:
@@ -864,7 +1013,9 @@ class SmartSearchTools:
                 ],
             }
 
-    def _search_in_dict(self, data: dict[str, Any] | list[Any] | Any, query: str) -> int:
+    def _search_in_dict(
+        self, data: dict[str, Any] | list[Any] | Any, query: str
+    ) -> int:
         """
         Recursively search for query string in nested dictionary/list structures.
 
