@@ -31,7 +31,7 @@ CATEGORY_DISPLAY = {v: k for k, v in CATEGORY_MAP.items()}
 CATEGORY_DISPLAY["plugin"] = "lovelace"  # Display as lovelace for users
 
 
-async def _check_hacs_available(client) -> tuple[bool, str | None]:
+async def _check_hacs_available(client: Any) -> tuple[bool, str | None]:
     """
     Check if HACS is installed and available via WebSocket.
 
@@ -56,7 +56,7 @@ async def _check_hacs_available(client) -> tuple[bool, str | None]:
         return False, f"Failed to connect to HACS: {str(e)}"
 
 
-def register_hacs_tools(mcp, client, **kwargs):
+def register_hacs_tools(mcp: Any, client: Any, **kwargs: Any) -> None:
     """Register HACS integration tools with the MCP server."""
 
     @mcp.tool(annotations={"idempotentHint": True, "readOnlyHint": True, "tags": ["hacs", "info"], "title": "Get HACS Info"})
@@ -104,6 +104,7 @@ def register_hacs_tools(mcp, client, **kwargs):
                 exception_to_structured_error(
                     Exception(f"HACS info request failed: {response}"),
                     context={"command": "hacs/info"},
+                    raise_error=True,
                 )
 
             result = response.get("result", {})
@@ -130,7 +131,8 @@ def register_hacs_tools(mcp, client, **kwargs):
                     "Check Home Assistant connection",
                     "Restart Home Assistant if HACS was recently installed",
                 ]
-            raise_tool_error(error_response)
+            error_with_tz = await add_timezone_metadata(client, error_response)
+            raise_tool_error(error_with_tz)
 
     @mcp.tool(annotations={"idempotentHint": True, "readOnlyHint": True, "tags": ["hacs", "search"], "title": "List HACS Installed"})
     @log_tool_usage
@@ -204,6 +206,7 @@ def register_hacs_tools(mcp, client, **kwargs):
                 exception_to_structured_error(
                     Exception(f"HACS repositories list request failed: {response}"),
                     context={"command": "hacs/repositories/list", "category": category},
+                    raise_error=True,
                 )
 
             repositories = response.get("result", [])
@@ -248,7 +251,8 @@ def register_hacs_tools(mcp, client, **kwargs):
                     "Check category name is valid: integration, lovelace, theme, appdaemon, python_script",
                     "Check Home Assistant connection",
                 ]
-            raise_tool_error(error_response)
+            error_with_tz = await add_timezone_metadata(client, error_response)
+            raise_tool_error(error_with_tz)
 
     @mcp.tool(annotations={"idempotentHint": True, "readOnlyHint": True, "tags": ["hacs", "search"], "title": "Search HACS Store"})
     @log_tool_usage
@@ -268,6 +272,13 @@ def register_hacs_tools(mcp, client, **kwargs):
                 description="Maximum number of results to return (default: 10, max: 100)",
             ),
         ] = 10,
+        offset: Annotated[
+            int | str,
+            Field(
+                default=0,
+                description="Number of results to skip for pagination (default: 0)",
+            ),
+        ] = 0,
     ) -> dict[str, Any]:
         """Search HACS store for repositories by keyword with pagination.
 
@@ -293,12 +304,13 @@ def register_hacs_tools(mcp, client, **kwargs):
             query: Search query (repository name, description, author)
             category: Filter by category (optional)
             max_results: Maximum results to return (default: 10, max: 100)
+            offset: Number of results to skip for pagination (default: 0)
 
         Returns:
             Search results from HACS store or error if HACS is not available.
         """
         try:
-            # Coerce max_results to int
+            # Coerce max_results and offset to int
             max_results_int = coerce_int_param(
                 max_results,
                 "max_results",
@@ -306,6 +318,12 @@ def register_hacs_tools(mcp, client, **kwargs):
                 min_value=1,
                 max_value=100,
             ) or 10
+            offset_int = coerce_int_param(
+                offset,
+                "offset",
+                default=0,
+                min_value=0,
+            ) or 0
 
             # Check if HACS is available
             is_available, error_msg = await _check_hacs_available(client)
@@ -337,6 +355,7 @@ def register_hacs_tools(mcp, client, **kwargs):
                 exception_to_structured_error(
                     Exception(f"HACS search request failed: {response}"),
                     context={"command": "hacs/repositories/list", "query": query, "category": category},
+                    raise_error=True,
                 )
 
             all_repositories = response.get("result", [])
@@ -383,16 +402,21 @@ def register_hacs_tools(mcp, client, **kwargs):
                         "score": score,
                     })
 
-            # Sort by score (descending) and limit results
+            # Sort by score (descending) and apply offset + limit
             matches.sort(key=lambda x: x["score"], reverse=True)
-            limited_matches = matches[:max_results_int]
+            limited_matches = matches[offset_int:offset_int + max_results_int]
+            has_more = (offset_int + len(limited_matches)) < len(matches)
 
             return await add_timezone_metadata(client, {
                 "success": True,
                 "query": query,
                 "category_filter": category,
                 "total_matches": len(matches),
-                "results_returned": len(limited_matches),
+                "offset": offset_int,
+                "limit": max_results_int,
+                "count": len(limited_matches),
+                "has_more": has_more,
+                "next_offset": offset_int + max_results_int if has_more else None,
                 "results": limited_matches,
             })
 
@@ -408,7 +432,8 @@ def register_hacs_tools(mcp, client, **kwargs):
                     "Try a simpler search query",
                     "Check category name is valid: integration, lovelace, theme, appdaemon, python_script",
                 ]
-            raise_tool_error(error_response)
+            error_with_tz = await add_timezone_metadata(client, error_response)
+            raise_tool_error(error_with_tz)
 
     @mcp.tool(annotations={"idempotentHint": True, "readOnlyHint": True, "tags": ["hacs", "info"], "title": "Get HACS Repository Info"})
     @log_tool_usage
@@ -485,6 +510,7 @@ def register_hacs_tools(mcp, client, **kwargs):
                 exception_to_structured_error(
                     Exception(f"HACS repository info request failed: {response}"),
                     context={"command": "hacs/repository/info", "repository_id": repository_id},
+                    raise_error=True,
                 )
 
             result = response.get("result", {})
@@ -524,7 +550,8 @@ def register_hacs_tools(mcp, client, **kwargs):
                     "Check repository ID format (e.g., 'hacs/integration' or 'owner/repo')",
                     "Use ha_hacs_search() to find the correct repository ID",
                 ]
-            raise_tool_error(error_response)
+            error_with_tz = await add_timezone_metadata(client, error_response)
+            raise_tool_error(error_with_tz)
 
     @mcp.tool(annotations={"destructiveHint": True, "tags": ["hacs", "management"], "title": "Add HACS Repository"})
     @log_tool_usage
@@ -618,6 +645,7 @@ def register_hacs_tools(mcp, client, **kwargs):
                         "repository": repository,
                         "category": category,
                     },
+                    raise_error=True,
                 )
 
             result = response.get("result", {})
@@ -649,7 +677,8 @@ def register_hacs_tools(mcp, client, **kwargs):
                     "Ensure category matches repository type",
                     "Check repository follows HACS guidelines: https://hacs.xyz/docs/publish/start",
                 ]
-            raise_tool_error(error_response)
+            error_with_tz = await add_timezone_metadata(client, error_response)
+            raise_tool_error(error_with_tz)
 
     @mcp.tool(annotations={"destructiveHint": True, "tags": ["hacs", "management"], "title": "Download/Install HACS Repository"})
     @log_tool_usage
@@ -753,6 +782,7 @@ def register_hacs_tools(mcp, client, **kwargs):
                         "repository_id": repository_id,
                         "version": version,
                     },
+                    raise_error=True,
                 )
 
             result = response.get("result", {})
@@ -784,4 +814,5 @@ def register_hacs_tools(mcp, client, **kwargs):
                     "Ensure the repository is in HACS (use ha_hacs_add_repository() if needed)",
                     "Check version format (e.g., 'v1.2.3' or '1.2.3')",
                 ]
-            raise_tool_error(error_response)
+            error_with_tz = await add_timezone_metadata(client, error_response)
+            raise_tool_error(error_with_tz)
