@@ -16,6 +16,7 @@ production-level functionality and compatibility.
 Tests are designed for Docker Home Assistant test environment at localhost:8124.
 """
 
+import asyncio
 import json
 import logging
 import time
@@ -772,7 +773,7 @@ class TestScriptOrchestration:
 
             # VERIFY: All scripts created with correct modes
 
-            for script_id, script_entity, expected_mode in created_scripts:
+            for script_id, _script_entity, expected_mode in created_scripts:
                 get_data = await mcp.call_tool_success(
                     "ha_config_get_script",
                     { "script_id": script_id}
@@ -799,7 +800,7 @@ class TestScriptOrchestration:
 
             # EXECUTE: Test each mode with timeout protection
             execution_tasks = []
-            for script_id, script_entity, mode in created_scripts:
+            for _script_id, script_entity, mode in created_scripts:
                 logger.info(f"🚀 Testing execution of {mode} mode script...")
 
                 # Execute the script
@@ -1247,7 +1248,7 @@ async def test_blueprint_script_lifecycle(
         logger.info(f"✅ Created blueprint script: {script_id}")
 
         # Step 4: Wait for script to be registered, then verify no sequence field
-        time.sleep(wait_for_script_registration())
+        await asyncio.sleep(wait_for_script_registration())
         get_result = await mcp.call_tool_success(
             "ha_config_get_script",
             {"script_id": script_id},
