@@ -1,15 +1,14 @@
 """Unit tests for voice assistant tools module."""
 
 import json
-
-import pytest
 from unittest.mock import AsyncMock, MagicMock
 
+import pytest
 from fastmcp.exceptions import ToolError
 
 from ha_mcp.tools.tools_voice_assistant import (
-    register_voice_assistant_tools,
     KNOWN_ASSISTANTS,
+    register_voice_assistant_tools,
 )
 
 
