@@ -45,8 +45,8 @@ class HomeAssistantTestEnvironment:
 
     def _setup_config_directory(self) -> Path:
         """Set up Home Assistant configuration directory."""
-        import tempfile
         import shutil
+        import tempfile
 
         # Create temporary directory for HA config
         config_dir = Path(tempfile.mkdtemp(prefix="ha_test_env_"))
@@ -103,7 +103,8 @@ class HomeAssistantTestEnvironment:
 
         # Create container with port configuration
         # renovate: datasource=docker depName=ghcr.io/home-assistant/home-assistant
-        container = DockerContainer("ghcr.io/home-assistant/home-assistant:2026.1.3")
+        HA_IMAGE = "ghcr.io/home-assistant/home-assistant:2026.1.3"
+        container = DockerContainer(HA_IMAGE)
 
         # Check for custom port via environment variable
         custom_port = os.environ.get("HA_TEST_PORT")
@@ -254,7 +255,7 @@ class HomeAssistantTestEnvironment:
                     print("✅ Ready")
                 else:
                     print(f"⚠️ Status {response.status_code}")
-            except:
+            except Exception:
                 print("❌ Not accessible")
         else:
             print("🐳 Container Status: Not running")
